src/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
dataclasses-json==0.5.14
Flask==2.3.3
frozenlist==1.4.0
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-spec==0.2.4
jsonschema-specifications==2023.7.1
langchain==0.0.277
langsmith==0.0.30
lazy-object-proxy==1.9.0
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.5
numpy==1.25.2
openai==0.28.0
openapi-schema-validator==0.6.0
openapi-spec-validator==0.6.0
packaging==23.1
pathable==0.4.3
prance==23.6.21.0
pydantic==2.3.0
pydantic_core==2.6.3
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rfc3339-validator==0.1.4
rfc3987==1.3.8
rpds-py==0.10.0
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
six==1.16.0
SQLAlchemy==2.0.20
strict-rfc3339==0.7
swagger-spec-validator==3.0.3
tenacity==8.2.3
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==2.0.4
validate-email==1.3
Werkzeug==2.3.7
yarl==1.9.2
